Revert "Implement VolToDb with audio_utils_power_from_amplitude"
This reverts commit 9a0d5e016155c8be8b0d474d26602bf89914fc14.

Reason for revert:Potential culprit for b/288445440 - verifying through ABTD before revert submission. This is part of the standard investigation process, and does not mean your CL will be reverted.

+++ media/libeffects/lvm/wrapper/Aidl/EffectBundleAidl.cpp
@@ -355,7 +355,7 @@ ndk::ScopedAStatus EffectBundleAidl::getParameterVolume(const Volume::Id& id,
     auto tag = id.get<Volume::Id::commonTag>();
     switch (tag) {
         case Volume::levelDb: {
-            volParam.set<Volume::levelDb>(static_cast<int>(mContext->getVolumeLevel()));
+            volParam.set<Volume::levelDb>(mContext->getVolumeLevel());
             break;
         }
         case Volume::mute: {
